+++ ardal/__init__.py
@@ -1,5 +1,4 @@
 from . import _ardal
 from .Ardal import Ardal
-from .ArdalParser import ArdalParser
 
-__all__ = ["Ardal", "ArdalParser"]
+__all__ = ["Ardal"]
