+++ src/HybridMatrix.hpp
@@ -18,9 +18,9 @@ public:
                  double density_threshold = 0.02);
 
     // distance functions: BitMatrix and RoaringMatrix
-    py::array_t<int> hamming( bool use_simd = true,
-                              int threads = 1,
-                              bool force_bit_backend = false ) const;
+    py::array_t<uint32_t> hamming( bool use_simd = true,
+                                   int threads = 1,
+                                   bool force_bit_backend = false ) const;
 
     py::array_t<int> innerProduct( bool use_simd = true,
                                    int threads = 1,
